+++ js/map.js
@@ -1,8 +1,22 @@
+import {showPopup} from './popup.js';
+
+const MARKER_ICON = {
+  iconUrl: '../img/main-pin.svg',
+  iconSize: [25, 41],
+  iconAnchor: [12.5, 41],
+};
+
+const MAP_LOCATION = {
+  lat: 35.6894,
+  lng: 139.692,
+};
+
+const MAP_SCALE = 10;
+
+const DISABLED_CLASS = 'map__filters--disabled';
+
 const map = L.map('map-canvas')
-  .setView({
-    lat: 35.6894,
-    lng: 139.692,
-  }, 10);
+  .setView(MAP_LOCATION, MAP_SCALE);
 
 L.tileLayer(
   'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
@@ -11,18 +25,44 @@ L.tileLayer(
   },
 ).addTo(map);
 
-export const markerGroup = L.layerGroup().addTo(map);
-export const setMapFilterEnabled = (enabled) => {
+const mapLoad = L.map('map-canvas').on('load');
+
+const markerGroup = L.layerGroup().addTo(map);
+
+export const createMarker = (offer) => {
+  const {lat, lng} = offer;
+
+  const icon = L.icon(MARKER_ICON);
+
+  const marker = L.marker(
+    {
+      lat,
+      lng,
+    },
+    {
+      icon,
+    },
+  );
+
+  marker
+    .addTo(markerGroup)
+    .bindPopup(
+      showPopup(offer),
+      {
+        keepInView: true,
+      },
+    );
+};
+
+export const setMapFilterEnabled = () => {
   const mapFilterDisabled = document.querySelector('.map__filters');
-  if (enabled) {
-    L.map('map-canvas').on('load', () => {
-      mapFilterDisabled.classList.remove('map__filters--disabled');
-      mapFilterDisabled.querySelectorAll('select').forEach((selectFilter) => {
-        selectFilter.classList.remove('disabled');
-      });
+  if (mapLoad) {
+    mapFilterDisabled.classList.remove(DISABLED_CLASS);
+    mapFilterDisabled.querySelectorAll('select').forEach((selectFilter) => {
+      selectFilter.classList.remove('disabled');
     });
   } else {
-    mapFilterDisabled.classList.add('map__filters--disabled');
+    mapFilterDisabled.classList.add(DISABLED_CLASS);
     mapFilterDisabled.querySelectorAll('select').forEach((selectFilter) => {
       selectFilter.classList.add('disabled');
     });
